# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a chatting API",
        "title": "Jackey Chatting API",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "version": "1"
    },
    "host": "localhost:8888",
    "basePath": "/api/",
    "paths": {
...
        "/message": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get messages between a user and a friend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Get messages",
                "parameters": [
                    {
                        "description": "Peer ID, offset and  limit",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/message.GetMessagesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler._BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/message.GetMessagesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/message/send": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send a message to a friend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Send a message",
                "parameters": [
                    {
                        "description": "Receiver ID and content",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/message.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Login a user with a username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "login request",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler._BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "description": "Logout a user by invalidating their token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Logout a user",
                "parameters": [
                    {
                        "description": "Token",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler._BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.LogoutResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
...
        "/user/register": {
            "post": {
                "description": "Register a new user with a username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "register request",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler._BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.RegisterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler._BaseResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据"
                },
                "status": {
                    "description": "状态",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "handler._ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "err msg"
                },
                "status": {
                    "description": "状态",
                    "type": "string",
                    "example": "failed"
                }
            }
        },
...
        "message.GetMessagesRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "每次请求返回的最大消息数",
                    "type": "integer"
                },
                "offset": {
                    "description": "从第几条记录开始读取",
                    "type": "integer"
                },
                "peer_id": {
                    "description": "聊天对象的用户ID",
                    "type": "integer"
                },
                "user_id": {
                    "description": "发起请求的用户ID",
                    "type": "integer"
                }
            }
        },
        "message.GetMessagesResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "description": "返回的消息列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/message.Message"
                    }
                }
            }
        },
        "message.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "msg_id": {
                    "type": "integer"
                },
                "receiver_id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "message.SendMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                }
            }
        },
...
        "user.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.LogoutResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "user.RegisterRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.RegisterResponse": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}